smsNotify/requirements.txt
# Used Python SDK
azure-identity
azure-storage-queue